Update Terraform recipe download error to include module version (#7620)
Updating Terraform recipe download error to include module version + minor refactoring (removed unnecessary wrapper function for tf.get, abstracting module related operations out of execute.go).
